+++ react/components/SKUSelector/components/SelectVariationMode.tsx
@@ -1,14 +1,21 @@
 import React from 'react'
 import { Dropdown } from 'vtex.styleguide'
-import { injectIntl, defineMessages, IntlShape } from 'react-intl'
+import type { IntlShape } from 'react-intl'
+import { injectIntl, defineMessages } from 'react-intl'
 
 import styles from '../styles.css'
-import { DisplayOption } from '../types'
+import type { DisplayOption } from '../types'
 
 interface VariationSelectModeProps {
   intl: IntlShape
   selectedItem: string | null
   displayOptions: DisplayOption[]
+}
+
+interface IsDisableParams {
+  disabled: boolean
+  impossible: boolean
+  available: boolean
 }
 
 const messages = defineMessages({
@@ -18,12 +25,26 @@ const messages = defineMessages({
   },
 })
 
+function isDisabled({ disabled, impossible, available }: IsDisableParams) {
+  if (disabled) {
+    if (impossible) return true
+    if (!impossible && !available) return true
+  }
+
+  return false
+}
+
 function SelectVariationMode(props: VariationSelectModeProps) {
   const { intl, selectedItem, displayOptions } = props
 
   const options = displayOptions.map(op => ({
     label: op.label,
     value: op.label,
+    disabled: isDisabled({
+      disabled: op.disabled,
+      impossible: op.impossible,
+      available: op.available,
+    }),
   }))
 
   const handleClick = (_: React.MouseEvent, value: string) => {